+++ src/types.d.ts
@@ -1,0 +1,11 @@
+interface Post {
+  excerpt?: string
+  body: string
+  frontmatter: {
+    title: string
+    date: Date
+    slug: string
+    language: string
+    tags: string[]
+  }
+}
